genet-node: remove warnings on windows

+++ genet-node/src/attr.cpp
@@ -76,7 +76,7 @@ NAN_GETTER(AttrWrapper::value) {
   v8::Local<v8::Value> val = Nan::Null();
   switch (var.type) {
   case Error:
-    val = Nan::Error(Nan::New(var.data, var.value.u64).ToLocalChecked());
+    val = Nan::Error(Nan::New(var.data, static_cast<int>(var.value.u64)).ToLocalChecked());
     genet_str_free(var.data);
     return;
   case Bool:
@@ -92,11 +92,11 @@ NAN_GETTER(AttrWrapper::value) {
     val = Nan::New(var.value.f64);
     break;
   case String:
-    val = Nan::New(var.data, var.value.u64).ToLocalChecked();
+    val = Nan::New(var.data, static_cast<int>(var.value.u64)).ToLocalChecked();
     genet_str_free(var.data);
     break;
   case Buffer:
-    val = Nan::CopyBuffer(var.data, var.value.u64).ToLocalChecked();
+    val = Nan::CopyBuffer(var.data, static_cast<uint32_t>(var.value.u64)).ToLocalChecked();
     genet_str_free(var.data);
     break;
   case ByteSlice: {
